fix: Businesses

+++ Assets/Scripts/Configs/ClickConfig.cs
@@ -5,6 +5,6 @@ namespace Configs
     [CreateAssetMenu(fileName = "ClickConfig", menuName = "Configs/ClickConfig")]
     public class ClickConfig: ScriptableObject
     {
-        public int TapAmount=10;
+        public long TapAmount=10;
     }
 }
