+++ Assets/Scripts/Configs/InvestitionConfig.cs
@@ -1,4 +1,3 @@
-using Gameplay.Estates.Generic;
 using Gameplay.Investitions;
 using UnityEngine;
 
@@ -9,9 +8,9 @@ namespace Configs
     {
         public InvestitionType Type;
         public string Name;
-        public int MinCost;
-        public int MaxCost;
-        public int InitialCost;
+        public long MinCost;
+        public long MaxCost;
+        public long InitialCost;
         public float ResumptionTime;
         public int MaxAmount;
         public float MeanReversionStrength;
